requirements.txt
coreapi==1.32.0
coverage==4.2
dj-database-url==0.4.1
Django==1.9.7
django-bootstrap3==7.0.1
django-filter==0.13.0
django-nose==1.4.4
django-rest-swagger==0.3.10
djangorestframework==3.4.1
djangorestframework-jwt==1.8.0
drf-extensions==0.2.8
drf-nested-routers==0.11.1
gunicorn==19.6.0
itypes==1.1.0
nose==1.3.7
openapi-codec==1.0.0
psycopg2==2.6.1
PyJWT==1.4.1
PyYAML==3.11
requests==2.10.0
selenium==2.53.6
simplejson==3.8.2
uritemplate==0.6
